remove unchecked calls in test code
