+++ testing/src/main/java/com/yubico/yubikit/testing/TestState.java
@@ -16,39 +16,31 @@
 
 package com.yubico.yubikit.testing;
 
-import static org.junit.Assume.assumeTrue;
-
 import com.yubico.yubikit.core.Transport;
 import com.yubico.yubikit.core.YubiKeyConnection;
 import com.yubico.yubikit.core.YubiKeyDevice;
-import com.yubico.yubikit.core.application.ApplicationNotAvailableException;
 import com.yubico.yubikit.core.application.ApplicationSession;
 import com.yubico.yubikit.core.application.CommandException;
 import com.yubico.yubikit.core.fido.FidoConnection;
-import com.yubico.yubikit.core.smartcard.ApduException;
 import com.yubico.yubikit.core.smartcard.SmartCardConnection;
 import com.yubico.yubikit.core.smartcard.scp.ScpKeyParams;
-import com.yubico.yubikit.core.smartcard.scp.SecurityDomainSession;
-import com.yubico.yubikit.fido.ctap.Ctap2Session;
 import com.yubico.yubikit.management.Capability;
 import com.yubico.yubikit.management.DeviceInfo;
 import com.yubico.yubikit.management.ManagementSession;
-import com.yubico.yubikit.oath.OathSession;
-import com.yubico.yubikit.openpgp.OpenPgpSession;
-import com.yubico.yubikit.piv.PivSession;
-import com.yubico.yubikit.testing.sd.SecurityDomainTestState;
 
 import java.io.IOException;
 
 import javax.annotation.Nullable;
 
 public class TestState {
-    public static class Builder<T extends Builder<T>> {
+    public abstract static class Builder<T extends Builder<T>> {
         final protected YubiKeyDevice device;
         @Nullable
         private Byte scpKid = null;
         @Nullable
         private ReconnectDeviceCallback reconnectDeviceCallback = null;
+
+        public abstract T getThis();
 
         public Builder(YubiKeyDevice device) {
             this.device = device;
@@ -56,19 +48,15 @@ public class TestState {
 
         public T scpKid(@Nullable Byte scpKid) {
             this.scpKid = scpKid;
-            //noinspection unchecked
-            return (T) this;
+            return getThis();
         }
 
         public T reconnectDeviceCallback(@Nullable ReconnectDeviceCallback reconnectDeviceCallback) {
             this.reconnectDeviceCallback = reconnectDeviceCallback;
-            //noinspection unchecked
-            return (T) this;
+            return getThis();
         }
 
-        public TestState build() throws Throwable {
-            return new TestState(this);
-        }
+        public abstract TestState build() throws Throwable;
     }
 
     protected YubiKeyDevice currentDevice;
@@ -91,11 +79,6 @@ public class TestState {
         return isUsbTransport;
     }
 
-    @SuppressWarnings("unused")
-    public interface DeviceCallback {
-        void invoke() throws Throwable;
-    }
-
     public interface StatefulDeviceCallback<S extends TestState> {
         void invoke(S state) throws Throwable;
     }
@@ -110,10 +93,6 @@ public class TestState {
 
     public interface SessionCallbackT<T extends ApplicationSession<T>, R> {
         R invoke(T session) throws Throwable;
-    }
-
-    public interface StatefulSessionCallbackT<T extends ApplicationSession<T>, S extends TestState, R> {
-        R invoke(T session, S state) throws Throwable;
     }
 
     public interface ReconnectDeviceCallback {
@@ -143,191 +122,6 @@ public class TestState {
     public boolean isFipsApproved(DeviceInfo deviceInfo, Capability capability) {
         return deviceInfo != null &&
                 (deviceInfo.getFipsApproved() & capability.bit) == capability.bit;
-    }
-
-    // PIV helpers
-    public <T extends TestState> void withPiv(StatefulSessionCallback<PivSession, T> callback)
-            throws Throwable {
-        try (SmartCardConnection connection = openSmartCardConnection()) {
-            final PivSession piv = getPivSession(connection, scpParameters);
-            assumeTrue("No PIV support", piv != null);
-            //noinspection unchecked
-            callback.invoke(piv, (T) this);
-        }
-        reconnect();
-    }
-
-    @Nullable
-    protected PivSession getPivSession(SmartCardConnection connection, ScpParameters scpParameters)
-            throws IOException {
-        try {
-            return new PivSession(connection, scpParameters.getKeyParams());
-        } catch (ApplicationNotAvailableException | ApduException ignored) {
-            // no PIV support
-        }
-        return null;
-    }
-
-    // CTAP2 helpers
-    public <R> R withCtap2(SessionCallbackT<Ctap2Session, R> callback) throws Throwable {
-        R result;
-        try (YubiKeyConnection connection = openConnection()) {
-            final Ctap2Session ctap2 = getCtap2Session(connection);
-            assumeTrue("No CTAP2 support", ctap2 != null);
-            result = callback.invoke(ctap2);
-        }
-        reconnect();
-        return result;
-    }
-
-    public void withCtap2(SessionCallback<Ctap2Session> callback) throws Throwable {
-        try (YubiKeyConnection connection = openConnection()) {
-            final Ctap2Session ctap2 = getCtap2Session(connection);
-            assumeTrue("No CTAP2 support", ctap2 != null);
-            callback.invoke(ctap2);
-        }
-        reconnect();
-    }
-
-    public <T extends TestState> void withCtap2(StatefulSessionCallback<Ctap2Session, T> callback)
-            throws Throwable {
-        try (YubiKeyConnection connection = openConnection()) {
-            final Ctap2Session ctap2 = getCtap2Session(connection);
-            assumeTrue("No CTAP2 support", ctap2 != null);
-            //noinspection unchecked
-            callback.invoke(ctap2, (T) this);
-        }
-        reconnect();
-    }
-
-    @Nullable
-    protected Ctap2Session getCtap2Session(YubiKeyConnection connection)
-            throws IOException, CommandException {
-        return (connection instanceof FidoConnection)
-                ? new Ctap2Session((FidoConnection) connection)
-                : connection instanceof SmartCardConnection
-                ? new Ctap2Session((SmartCardConnection) connection)
-                : null;
-    }
-
-    // OATH helpers
-    public void withOath(SessionCallback<OathSession> callback) throws Throwable {
-        try (SmartCardConnection connection = openSmartCardConnection()) {
-            callback.invoke(getOathSession(connection, scpParameters));
-        }
-        reconnect();
-    }
-
-    public <T extends TestState> void withOath(StatefulSessionCallback<OathSession, T> callback)
-            throws Throwable {
-        try (SmartCardConnection connection = openSmartCardConnection()) {
-            //noinspection unchecked
-            callback.invoke(getOathSession(connection, scpParameters), (T) this);
-        }
-        reconnect();
-    }
-
-    @Nullable
-    protected OathSession getOathSession(SmartCardConnection connection, ScpParameters scpParameters)
-            throws IOException {
-        try {
-            return new OathSession(connection, scpParameters.getKeyParams());
-        } catch (ApplicationNotAvailableException ignored) {
-            // no OATH support
-        }
-        return null;
-    }
-
-    // Security domain helpers
-    public void withSecurityDomain(SessionCallback<SecurityDomainSession> callback) throws Throwable {
-        try (SmartCardConnection connection = openSmartCardConnection()) {
-            callback.invoke(getSecurityDomainSession(connection));
-        }
-        reconnect();
-    }
-
-    public <R> R withSecurityDomain(SessionCallbackT<SecurityDomainSession, R> callback) throws Throwable {
-        R result;
-        try (SmartCardConnection connection = openSmartCardConnection()) {
-            result = callback.invoke(getSecurityDomainSession(connection));
-        }
-        reconnect();
-        return result;
-    }
-
-    public void withSecurityDomain(ScpKeyParams scpKeyParams, SessionCallback<SecurityDomainSession> callback) throws Throwable {
-        try (SmartCardConnection connection = openSmartCardConnection()) {
-            callback.invoke(getSecurityDomainSession(scpKeyParams, connection));
-        }
-        reconnect();
-    }
-
-    public <T extends TestState> void withSecurityDomain(StatefulSessionCallback<SecurityDomainSession, T> callback)
-            throws Throwable {
-        try (SmartCardConnection connection = openSmartCardConnection()) {
-            //noinspection unchecked
-            callback.invoke(getSecurityDomainSession(connection), (T) this);
-        }
-        reconnect();
-    }
-
-    public <R> R withSecurityDomain(ScpKeyParams scpKeyParams, SessionCallbackT<SecurityDomainSession, R> callback) throws Throwable {
-        R result;
-        try (SmartCardConnection connection = openSmartCardConnection()) {
-            result = callback.invoke(getSecurityDomainSession(scpKeyParams, connection));
-        }
-        reconnect();
-        return result;
-    }
-
-    @Nullable
-    protected SecurityDomainSession getSecurityDomainSession(SmartCardConnection connection)
-            throws IOException {
-        try {
-            return new SecurityDomainSession(connection, scpParameters.getKeyParams());
-        } catch (ApplicationNotAvailableException ignored) {
-            // no OATH support
-        }
-        return null;
-    }
-
-    @Nullable
-    protected SecurityDomainSession getSecurityDomainSession(ScpKeyParams scpKeyParams, SmartCardConnection connection)
-            throws IOException {
-        try {
-            return new SecurityDomainSession(connection, scpKeyParams);
-        } catch (ApplicationNotAvailableException ignored) {
-            // no OATH support
-        }
-        return null;
-    }
-
-    // OpenPGP helpers
-    public <T extends TestState> void withOpenPgp(StatefulSessionCallback<OpenPgpSession, T> callback)
-            throws Throwable {
-        try (SmartCardConnection connection = openSmartCardConnection()) {
-            //noinspection unchecked
-            callback.invoke(getOpenPgpSession(connection, scpParameters), (T) this);
-        }
-        reconnect();
-    }
-
-    @Nullable
-    protected OpenPgpSession getOpenPgpSession(SmartCardConnection connection, ScpParameters scpParameters)
-            throws IOException, CommandException {
-        try {
-            return new OpenPgpSession(connection, scpParameters.getKeyParams());
-        } catch (ApplicationNotAvailableException ignored) {
-            // no OpenPgp support
-        }
-        return null;
-    }
-
-    // device helper
-    public <T extends TestState> void withDeviceCallback(StatefulDeviceCallback<T> callback)
-            throws Throwable {
-        //noinspection unchecked
-        callback.invoke((T) this);
     }
 
     // connection helpers
