+++ testing/src/main/java/com/yubico/yubikit/testing/openpgp/OpenPgpTestState.java
@@ -21,14 +21,21 @@ import static org.junit.Assert.assertTrue;
 import static org.junit.Assume.assumeTrue;
 
 import com.yubico.yubikit.core.YubiKeyDevice;
+import com.yubico.yubikit.core.application.ApplicationNotAvailableException;
+import com.yubico.yubikit.core.application.CommandException;
 import com.yubico.yubikit.core.smartcard.SmartCardConnection;
 import com.yubico.yubikit.management.Capability;
 import com.yubico.yubikit.management.DeviceInfo;
 import com.yubico.yubikit.openpgp.OpenPgpSession;
 import com.yubico.yubikit.openpgp.Pw;
+import com.yubico.yubikit.testing.ScpParameters;
 import com.yubico.yubikit.testing.TestState;
 
 import org.junit.Assume;
+
+import java.io.IOException;
+
+import javax.annotation.Nullable;
 
 public class OpenPgpTestState extends TestState {
 
@@ -43,6 +50,11 @@ public class OpenPgpTestState extends TestState {
 
         public Builder(YubiKeyDevice device) {
             super(device);
+        }
+
+        @Override
+        public Builder getThis() {
+            return this;
         }
 
         public OpenPgpTestState build() throws Throwable {
@@ -97,5 +109,24 @@ public class OpenPgpTestState extends TestState {
             assertNotNull(deviceInfo);
             assertTrue("Device not OpenPgp FIPS approved as expected", isFipsApproved);
         }
+    }
+
+    public void withOpenPgp(StatefulSessionCallback<OpenPgpSession, OpenPgpTestState> callback)
+            throws Throwable {
+        try (SmartCardConnection connection = openSmartCardConnection()) {
+            callback.invoke(getOpenPgpSession(connection, scpParameters), this);
+        }
+        reconnect();
+    }
+
+    @Nullable
+    public static OpenPgpSession getOpenPgpSession(SmartCardConnection connection, ScpParameters scpParameters)
+            throws IOException, CommandException {
+        try {
+            return new OpenPgpSession(connection, scpParameters.getKeyParams());
+        } catch (ApplicationNotAvailableException ignored) {
+            // no OpenPgp support
+        }
+        return null;
     }
 }
